# src/docs/swagger.yaml
definitions:
...
  dtos.CommentResponse:
    properties:
      dislikes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      message:
        type: string
      tweet:
        $ref: '#/definitions/dtos.TweetResponse'
      tweet_id:
        type: integer
      user:
        $ref: '#/definitions/dtos.UserResponse'
      user_id:
        type: integer
    type: object
...
  dtos.TweetCreate:
    properties:
      message:
        maxLength: 1000
        minLength: 10
        type: string
      title:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - message
    - title
    type: object
  dtos.TweetResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/dtos.CommentResponse'
        type: array
      dislikes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      message:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/dtos.UserResponse'
      user_id:
        type: integer
    type: object
  dtos.TweetUpdate:
    properties:
      message:
        maxLength: 1000
        minLength: 10
        type: string
      title:
        maxLength: 50
        minLength: 2
        type: string
    type: object
...
  dtos.UserResponse:
    properties:
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      mobile_number:
        type: string
      username:
        type: string
    type: object
...
  models.Comment:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      dislikes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      enabled:
        type: boolean
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      message:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      tweet:
        $ref: '#/definitions/models.Tweet'
      tweet_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Tweet:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      dislikes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      enabled:
        type: boolean
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      message:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      comment_dislikes:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comment_likes:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      enabled:
        type: boolean
      firstname:
        type: string
      followers:
        items:
          $ref: '#/definitions/models.User'
        type: array
      followings:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      lastname:
        type: string
      mobile_number:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      password:
        type: string
      tweet_dislikes:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      tweet_likes:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      tweets:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      username:
        type: string
    type: object
  responses.Response:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
      status:
        type: boolean
      statusCode:
        type: integer
      validationError:
        items:
          $ref: '#/definitions/validations.ValidationError'
        type: array
    type: object
  validations.ValidationError:
    properties:
      field:
        type: string
      param:
        type: string
      tag:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /tweet/delete:
    delete:
      description: delete tweet by id and its owner information
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Delete Tweet
      tags:
      - Tweet
  /tweet/dislike:
    post:
      description: dislike tweet
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Dislike Tweet
      tags:
      - Tweet
  /tweet/explore:
    get:
      description: delete tweet by id and its owner information
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/dtos.TweetResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Followings Tweets
      tags:
      - Tweet
  /tweet/get/followings/tweets:
    get:
      description: delete tweet by id and its owner information
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/dtos.TweetResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Followings Tweets
      tags:
      - Tweet
  /tweet/get/tweet:
    get:
      description: get tweet by id
      parameters:
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/models.Tweet'
              type: object
        "400":
          description: Validation Error
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Tweet
      tags:
      - Tweet
  /tweet/get/tweets:
    get:
      description: get tweets those tweeted by an user
      parameters:
      - description: get user's username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/dtos.TweetResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Tweets
      tags:
      - Tweet
  /tweet/like:
    post:
      description: like tweet
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Like Tweet
      tags:
      - Tweet
  /tweet/post:
    post:
      consumes:
      - application/json
      description: tweet post
      parameters:
      - description: get user's username
        in: query
        name: username
        required: true
        type: string
      - description: get user's password
        in: query
        name: password
        required: true
        type: string
      - description: get title and message of new tweet
        in: body
        name: TweetCreate
        required: true
        schema:
          $ref: '#/definitions/dtos.TweetCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/dtos.TweetResponse'
              type: object
        "400":
          description: Validation Error
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Post Tweet
      tags:
      - Tweet
  /tweet/update:
    put:
      consumes:
      - application/json
      description: update tweet's title and message
      parameters:
      - description: owner's username
        in: query
        name: username
        required: true
        type: string
      - description: owner's password
        in: query
        name: password
        required: true
        type: string
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      - description: update's fields
        in: body
        name: TweetUpdate
        required: true
        schema:
          $ref: '#/definitions/dtos.TweetUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/dtos.TweetResponse'
              type: object
        "400":
          description: Validation Error
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Update Tweet
      tags:
      - Tweet
...
swagger: "2.0"